pip_requirements.txt
lxml
matplotlib
numpy
opencv-python
pillow
pyflakes
pyyaml
scipy
textual
tqdm
